Introducing get_number_by_idx interface for retrieving literal number value by it's identifier. Change ecma_Number_t type from double to float.

+++ src/libecmaobjects/ecma-globals.h
@@ -332,7 +332,7 @@ typedef uint8_t ecma_Char_t;
 /**
  * Description of an ecma-number
  */
-typedef double ecma_Number_t;
+typedef float ecma_Number_t;
 
 /**
  * Description of arrays'/strings' length
